refactor: deconsolidate specific tests run individually in CI to prevent full test suite re-compilations
