+++ .github/workflows/ci.yml
@@ -65,15 +65,15 @@ jobs:
       - name: Run Docker-dependent tests
         if: runner.os == 'Linux'
         run: |
-          cargo test -p autumn-web --test integration_tests integration::db_hooks_lifecycle -- --ignored
-          cargo test -p autumn-web --test integration_tests integration::repository_shard_replica_routing -- --ignored
-          cargo test -p autumn-web --all-features --test integration_tests integration::feature_flags_pg_integration -- --ignored
-          cargo test -p autumn-web --test integration_tests integration::scoped_tokens_db -- --ignored
+          cargo test -p autumn-web --test db_hooks_lifecycle -- --ignored
+          cargo test -p autumn-web --test repository_shard_replica_routing -- --ignored
+          cargo test -p autumn-web --all-features --test feature_flags_pg_integration -- --ignored
+          cargo test -p autumn-web --test scoped_tokens_db -- --ignored
           cargo test -p autumn-admin-plugin --test token_admin_db -- --ignored
           cargo test -p autumn-cache-redis -- --ignored
-          cargo test -p autumn-cli --test cli_tests integration::flags -- --ignored
-          cargo test -p autumn-cli --test cli_tests integration::config -- --ignored
-          cargo test -p autumn-cli --test cli_tests integration::experiments -- --ignored
+          cargo test -p autumn-cli --test flags -- --ignored
+          cargo test -p autumn-cli --test config -- --ignored
+          cargo test -p autumn-cli --test experiments -- --ignored
 
   coverage:
     name: Coverage
@@ -105,19 +105,19 @@ jobs:
           cargo llvm-cov clean --workspace
           cargo llvm-cov --workspace --all-features --no-report
           cargo llvm-cov --no-report -p autumn-web --all-features \
-            --test integration_tests integration::db_hooks_lifecycle -- --ignored
+            --test db_hooks_lifecycle -- --ignored
           cargo llvm-cov --no-report -p autumn-web --all-features \
-            --test integration_tests integration::repository_shard_replica_routing -- --ignored
+            --test repository_shard_replica_routing -- --ignored
           cargo llvm-cov --no-report -p autumn-web --all-features \
-            --test integration_tests integration::feature_flags_pg_integration -- --ignored
+            --test feature_flags_pg_integration -- --ignored
           cargo llvm-cov --no-report -p autumn-web --all-features \
-            --test integration_tests integration::scoped_tokens_db -- --ignored
+            --test scoped_tokens_db -- --ignored
           cargo llvm-cov --no-report -p autumn-admin-plugin \
             --test token_admin_db -- --ignored
           cargo llvm-cov --no-report -p autumn-cache-redis -- --ignored
-          cargo llvm-cov --no-report -p autumn-cli --test cli_tests integration::flags -- --ignored
-          cargo llvm-cov --no-report -p autumn-cli --test cli_tests integration::config -- --ignored
-          cargo llvm-cov --no-report -p autumn-cli --test cli_tests integration::experiments -- --ignored
+          cargo llvm-cov --no-report -p autumn-cli --test flags -- --ignored
+          cargo llvm-cov --no-report -p autumn-cli --test config -- --ignored
+          cargo llvm-cov --no-report -p autumn-cli --test experiments -- --ignored
           cargo llvm-cov report --lcov --output-path lcov.info
       - name: Upload to Codecov
         uses: codecov/codecov-action@v6
