+++ autumn-cli/tests/cli_tests.rs
@@ -1,2 +1,3 @@
 //! Main entry point for consolidated CLI integration tests.
+mod common;
 mod integration;
