+++ autumn-cli/tests/generate.rs
