+++ autumn-cli/tests/integration/mod.rs
@@ -1,12 +1,6 @@
 mod cloud_native_scaffold;
-pub mod common;
-mod config;
 mod db;
 mod db_pull;
-mod e2e;
-mod experiments;
-mod flags;
-mod generate;
 mod generate_references_postgres;
 mod migrate_down;
 mod repo_hygiene;
